requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
cachetools==5.5.0
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
opuslib==3.0.1
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
yarl==1.18.3
youtube-dl==2021.12.17
yt-dlp==2024.12.23
